Redid HTML & CSS
Remade and edited my old HTML & CSS to make it more readable and
maintainable.

Though as a result, I have removed a lot of the ad-hoc scaffolding for
the Schedule implementation, though it wasn't really _that_ good anyways.

Other than that, it's pretty good! It's like 99% the same as it was in
the last commit.