Added boilerplate for the Settings page.
It's just the CSS and HTML to allow you to jump between the two pages
and not get stuck.

Though, I did notice that the bold-ing of the header link to show you're
there does actually change the size of it's container, causing them to
not align perfectly between pages. :P